fix(proxy): support legacy http proxy server http/1.0 response for CONNECT (#19)
There are some cases where the websocket connection may go through a legacy http proxy server which only supports HTTP/1.0. When doing CONNECT request, the http proxy server will respond with HTTP/1.0 200 and that may fail the status line check in the current code. This PR loosen this a bit to support both HTTP/1.1 and HTTP/1.0 response for the CONNECT request.

https://konghq.atlassian.net/browse/FTI-6683